refactor: move Validation to protocols folder
issue: #6
