+++ src/presentation/controllers/signup/signup-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../protocols'
 export * from '../../../domain/models/account'
 export * from '../../../domain/usecases/add-account'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
